refactor(vault): split orchestrators out of mod.rs into orchestrators.rs
Move `create_vault`, `open_vault`, `save_block`, `share_block` plus
their private helpers (`format_uuid_hyphenated`,
`ensure_empty_directory`, `build_owner_card_from_bundle`, `tick_clock`)
and the orchestrator unit tests from `vault/mod.rs` into a new
`vault/orchestrators.rs`.

mod.rs goes from 1600 LoC to 226 LoC; the self-imposed 600-LoC
threshold flagged in the prior doc comment was exceeded by 2.6×, and
the review on PR #5 deferred the split as a follow-up.

The public API is unchanged: the four functions plus `OpenVault` and
`Unlocker` are re-exported from `vault::mod` so callers continue to
import them as `secretary_core::vault::create_vault`, etc. The 345-test
workspace suite and `cargo clippy --all-targets --workspace -- -D
warnings` remain green.

Follow-up to PR #5 review.
